@@ -24,59 +24,50 @@
     "clean": "git clean -xfd -i -e node_modules ."
   },
   "devDependencies": {
-    "@types/classnames": "^2.2.9",
-    "@types/react": "^16.9.19",
-    "@types/react-helmet": "^5.0.15",
-    "@types/react-toggle": "^4.0.2",
-    "@types/recharts": "^1.8.5",
-    "@typescript-eslint/eslint-plugin": "^2.19.0",
-    "@typescript-eslint/parser": "^2.19.0",
-    "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.0.1",
-    "eslint-config-standard-react": "^9.0.0",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.1.1",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^2.3.0",
-    "eslint-plugin-simple-import-sort": "^5.0.1",
-    "eslint-plugin-standard": "^4.0.1",
-    "nodemon": "^2.0.2",
-    "stylelint": "^13.0.0",
-    "stylelint-config-rational-order": "^0.1.1",
-    "stylelint-config-standard": "^19.0.0"
-  },
-  "dependencies": {
-    "@react-pdf/renderer": "^1.6.8",
+    "@react-pdf/renderer": "^1.6.12",
     "@tippy.js/react": "^3.1.1",
+    "@types/classnames": "^2.2.11",
+    "@types/react": "^17.0.0",
+    "@types/react-helmet": "^6.1.0",
+    "@types/react-toggle": "^4.0.2",
+    "@types/recharts": "^1.8.18",
+    "autoprefixer": "^10.1.0",
     "classnames": "^2.2.6",
-    "evergreen-ui": "^4.23.0",
-    "gatsby": "^2.19.12",
-    "gatsby-plugin-fullstory": "^2.1.21",
-    "gatsby-plugin-gtag": "^1.0.12",
-    "gatsby-plugin-manifest": "^2.2.41",
-    "gatsby-plugin-netlify": "^2.1.32",
-    "gatsby-plugin-offline": "^3.0.34",
-    "gatsby-plugin-postcss": "^2.1.20",
-    "gatsby-plugin-react-helmet": "^3.1.22",
-    "gatsby-plugin-resolve-src": "^2.0.0",
-    "gatsby-plugin-typescript": "^2.1.27",
-    "globby": "^11.0.0",
-    "js-beautify": "^1.10.3",
+    "eslint": "^7.16.0",
+    "eslint-config-vectron": "^2.2.0",
+    "evergreen-ui": "^5.1.2",
+    "gatsby": "^2.29.1",
+    "gatsby-plugin-fullstory": "^2.7.0",
+    "gatsby-plugin-gtag": "^1.0.13",
+    "gatsby-plugin-manifest": "^2.9.0",
+    "gatsby-plugin-netlify": "^2.8.0",
+    "gatsby-plugin-offline": "^3.7.0",
+    "gatsby-plugin-postcss": "3.3.0",
+    "gatsby-plugin-react-helmet": "^3.7.0",
+    "gatsby-plugin-resolve-src": "^2.1.0",
+    "gatsby-plugin-typescript": "^2.9.0",
+    "globby": "^11.0.1",
+    "js-beautify": "^1.13.0",
     "module-alias": "^2.2.1",
-    "postcss-nested": "^4.2.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-flip-toolkit": "^7.0.7",
-    "react-helmet": "^5.2.1",
-    "react-icons": "^3.9.0",
-    "react-qr-svg": "^2.2.1",
+    "nodemon": "^2.0.6",
+    "postcss": "^8.2.1",
+    "postcss-nested": "^5.0.2",
+    "postcss-preset-env": "^6.7.0",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-flip-toolkit": "^7.0.13",
+    "react-helmet": "^6.1.0",
+    "react-icons": "^4.1.0",
+    "react-qr-svg": "^2.3.0",
     "react-toggle": "^4.0.2",
-    "recharts": "^1.7.1",
-    "tippy.js": "^5.2.0",
-    "ts-node": "^8.6.2",
-    "typescript": "~3.7.5",
+    "recharts": "^1.8.5",
+    "stylelint": "^13.8.0",
+    "stylelint-config-rational-order": "^0.1.1",
+    "stylelint-config-standard": "^20.0.0",
+    "tippy.js": "^6.2.7",
+    "ts-node": "^9.1.1",
+    "typescript": "~4.1.3",
+    "typescript-plugin-css-modules": "^3.0.1",
     "utility-types": "^3.10.0",
     "vcards-js": "^2.10.0"
   },
