@@ -1,67 +1,10 @@
 module.exports = {
   root: true,
   extends: [
-    'standard',
-    'standard-react',
-    'plugin:import/typescript'
+    'vectron',
   ],
-  env: {
-    browser: true,
-    node: true,
-  },
-  plugins: [
-    'react',
-    'react-hooks',
-    'import',
-    'simple-import-sort',
-  ],
-  overrides: [
-    {
-      files: ['**/*.ts', '**/*.tsx'],
-      parser: '@typescript-eslint/parser',
-      plugins: ['@typescript-eslint'],
-      rules: {
-        'no-unused-vars': 'off',
-        '@typescript-eslint/no-unused-vars': ['error', {
-          vars: 'all',
-          args: 'after-used',
-          ignoreRestSiblings: false,
-        }],
-      },
-    },
-  ],
-  settings: {
-    'import/resolver': {
-      typescript: {
-        directory: __dirname,
-      },
-      webpack: 'webpack.config.js',
-    },
-    react: {
-      version: 'detect',
-    },
-  },
   rules: {
-    'no-console': ['warn', { allow: ['warn', 'error'] }],
-    'quote-props': ['error', 'as-needed'],
-    'comma-dangle': ['warn', {
-      arrays: 'only-multiline',
-      objects: 'always-multiline',
-      imports: 'only-multiline',
-    }],
-    quotes: ['warn', 'single'],
-    'space-before-function-paren': ['error', {
-      anonymous: 'never',
-      named: 'never',
-      asyncArrow: 'always',
-    }],
-    'jsx-quotes': ['error', 'prefer-double'],
-    'import/first': 'error',
-    'import/newline-after-import': 'error',
-    'import/no-extraneous-dependencies': 1,
-    'import/no-unresolved': [0, { caseSensitive: true }],
-    'import/order': 'off',
-    'simple-import-sort/sort': ['error', {
+    'simple-import-sort/imports': ['error', {
       groups: [
         // Node.js builtins. You could also generate this regex if you use a `.js` config.
         // For example: `^(${require("module").builtinModules.join("|")})(/|$)`
@@ -71,7 +14,7 @@ module.exports = {
         // Packages. `react` related packages come first.
         ['^react', '^@?\\w'],
         // Internal packages.
-        ['^(@|components|exts|data|hooks|idb|layouts|logo|pages|types|utils|api|api_supplimental)(/.*|$)'],
+        ['^(@|components|hooks|data|contexts|layouts|pages|types|utils|api|api_supplimental)(/.*|$)'],
         // Side effect imports.
         ['^\\u0000'],
         // Parent imports. Put `..` last.
@@ -81,22 +24,6 @@ module.exports = {
         // Style imports.
         ['^.+\\.s?css$'],
       ],
-    }],
-    'sort-imports': 'off',
-    'react-hooks/rules-of-hooks': 'error',
-    'react/jsx-closing-tag-location': 2,
-    'react/jsx-closing-bracket-location': 2,
-    'react/jsx-indent': [2, 2, { checkAttributes: true }],
-    'react/jsx-max-props-per-line': ['error', { maximum: 1, when: 'multiline' }],
-    'react/prop-types': ['error', { ignore: ['children', 'className', 'style'] }],
-    'keyword-spacing': ['error', {
-      overrides: {
-        if: { after: false },
-        for: { after: false },
-        while: { after: false },
-        catch: { after: false },
-        switch: { after: false },
-      },
     }],
   },
 }
