@@ -25,7 +25,7 @@ export default function useLocalStorage<T>(key: string, initialValue?: T) {
   })
   // Return a wrapped version of useState's setter function that persists the
   // new value to localStorage.
-  const setValue: (value: (v: T) => T | T) => void = (value) => {
+  const setValue: (_value: (_v: T) => T | T) => void = (value) => {
     if(typeof window === 'undefined') return
     try {
       // Allow value to be a function so we have same API as useState
