@@ -4,7 +4,7 @@ import AllSkills from 'components/AllSkills'
 import Jobs from 'components/Jobs'
 import Layout from 'components/Layout'
 import allJobs from 'data/jobs'
-import skillsById from 'data/skills'
+import { skills as skillsById } from 'data/skills'
 
 const skills = Object.values(skillsById)
 export default class IndexPage extends React.Component {
